Adding a resizeable screen class, as most classes use the same functions, hopefuly to reduce code duplication in this package. Also optimises imports for the whole package.
# Conflicts:
#	source/core/src/main/com/csse3200/game/screens/CombatScreen.java
