+++ source/core/src/main/com/csse3200/game/screens/BirdieDashScreen.java
@@ -1,5 +1,9 @@
 package com.csse3200.game.screens;
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.Stage;
@@ -7,31 +11,27 @@ import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
-import com.badlogic.gdx.Screen;
-import com.csse3200.game.gamestate.GameState;
-import com.csse3200.game.minigames.KeyboardMiniGameInputComponent;
-import com.csse3200.game.minigames.birdieDash.BirdieDashGame;
-import com.csse3200.game.minigames.birdieDash.controller.KeyboardBirdInputComponent;
-import com.csse3200.game.input.InputComponent;
-import com.csse3200.game.input.InputDecorator;
-import com.csse3200.game.rendering.Renderer;
-import com.csse3200.game.services.AudioManager;
-import com.csse3200.game.services.ServiceContainer;
-import com.csse3200.game.ui.minigames.ScoreBoard;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.GL20;
 import com.csse3200.game.GdxGame;
+import com.csse3200.game.components.gamearea.PerformanceDisplay;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.EntityService;
 import com.csse3200.game.entities.factories.RenderFactory;
+import com.csse3200.game.gamestate.GameState;
+import com.csse3200.game.input.InputComponent;
+import com.csse3200.game.input.InputDecorator;
 import com.csse3200.game.input.InputService;
+import com.csse3200.game.minigames.KeyboardMiniGameInputComponent;
+import com.csse3200.game.minigames.birdieDash.BirdieDashGame;
+import com.csse3200.game.minigames.birdieDash.controller.KeyboardBirdInputComponent;
 import com.csse3200.game.rendering.RenderService;
+import com.csse3200.game.rendering.Renderer;
+import com.csse3200.game.services.AudioManager;
 import com.csse3200.game.services.GameTime;
+import com.csse3200.game.services.ServiceContainer;
 import com.csse3200.game.services.ServiceLocator;
-import com.csse3200.game.components.gamearea.PerformanceDisplay;
+import com.csse3200.game.ui.minigames.ScoreBoard;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import static com.csse3200.game.minigames.MiniGameNames.BIRD;
