+++ source/core/src/main/com/csse3200/game/screens/BossCutsceneScreen.java
@@ -2,7 +2,6 @@ package com.csse3200.game.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.ScreenAdapter;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
@@ -16,13 +15,9 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.csse3200.game.GdxGame;
 import com.csse3200.game.components.CombatStatsComponent;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.EntityService;
-import com.csse3200.game.entities.factories.RenderFactory;
 import com.csse3200.game.input.InputService;
 import com.csse3200.game.physics.PhysicsEngine;
 import com.csse3200.game.physics.PhysicsService;
-import com.csse3200.game.rendering.RenderService;
-import com.csse3200.game.rendering.Renderer;
 import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceContainer;
@@ -34,7 +29,7 @@ import org.slf4j.LoggerFactory;
  * Manages the cutscene for Boss NPCs displayed before transitioning to the combat screen.
  * Handles initialization, rendering, and disposal of cutscene elements.
  */
-public class BossCutsceneScreen extends ScreenAdapter {
+public class BossCutsceneScreen extends ResizableScreen {
     private static final float CUTSCENE_DURATION = 5.0f; // Cutscene lasts for 3 seconds
     private float timeElapsed = 0;
     private boolean transition;
@@ -43,7 +38,6 @@ public class BossCutsceneScreen extends ScreenAdapter {
     private static final Vector2 CAMERA_POSITION = new Vector2(7.5f, 7.5f);
     private boolean isPaused = false;
     private final GdxGame game;
-    private final Renderer renderer;
     private final PhysicsEngine physicsEngine;
     private final Screen oldScreen;
     private final ServiceContainer oldScreenServices;
@@ -60,12 +54,14 @@ public class BossCutsceneScreen extends ScreenAdapter {
      * @param enemy the enemy entity
      */
     public BossCutsceneScreen(GdxGame game, Screen screen, ServiceContainer container, Entity player, Entity enemy) {
+        super();
+        
         this.game = game;
         this.oldScreen = screen;
         this.oldScreenServices = container;
         this.player = player;
         this.enemy = enemy;
-
+        
         logger.debug("Initializing boss cutscene screen services");
         ServiceLocator.registerTimeSource(new GameTime());
 
@@ -75,11 +71,7 @@ public class BossCutsceneScreen extends ScreenAdapter {
 
         ServiceLocator.registerInputService(new InputService());
         ServiceLocator.registerResourceService(new ResourceService());
-
-        ServiceLocator.registerEntityService(new EntityService());
-        ServiceLocator.registerRenderService(new RenderService());
-
-        renderer = RenderFactory.createRenderer();
+        
         renderer.getCamera().getEntity().setPosition(CAMERA_POSITION);
         renderer.getDebug().renderPhysicsWorld(physicsEngine.getWorld());
 
@@ -104,13 +96,7 @@ public class BossCutsceneScreen extends ScreenAdapter {
             }
         }
     }
-
-    @Override
-    public void resize(int width, int height) {
-        renderer.resize(width, height);
-        logger.trace("Resized renderer: ({} x {})", width, height);
-    }
-
+    
     @Override
     public void pause() {
         isPaused = true;
@@ -126,14 +112,8 @@ public class BossCutsceneScreen extends ScreenAdapter {
     @Override
     public void dispose() {
         logger.debug("Disposing cutscene screen");
-
-        renderer.dispose();
-
-        ServiceLocator.getEntityService().dispose();
-        ServiceLocator.getRenderService().dispose();
         ServiceLocator.getResourceService().dispose();
-
-        ServiceLocator.clear();
+        super.dispose();
     }
 
     /**
