+++ source/core/src/main/com/csse3200/game/screens/EndMiniGameScreen.java
@@ -7,29 +7,29 @@ import com.badlogic.gdx.ScreenAdapter;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
-import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
-import com.badlogic.gdx.scenes.scene2d.ui.Table;
+import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
+import com.badlogic.gdx.scenes.scene2d.ui.Table;
+import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.csse3200.game.GdxGame;
 import com.csse3200.game.areas.MapHandler;
+import com.csse3200.game.components.inventory.PlayerInventoryDisplay;
+import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.factories.LootBoxFactory;
 import com.csse3200.game.inventory.items.lootbox.UniversalLootBox;
 import com.csse3200.game.minigames.MiniGameConstants;
 import com.csse3200.game.minigames.MiniGameMedals;
 import com.csse3200.game.minigames.MiniGameNames;
-import com.csse3200.game.components.inventory.PlayerInventoryDisplay;
-import com.csse3200.game.entities.Entity;
 import com.csse3200.game.services.AudioManager;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceContainer;
 import com.csse3200.game.services.ServiceLocator;
-import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import static com.csse3200.game.minigames.MiniGameNames.BIRD;
 import static com.csse3200.game.minigames.MiniGameNames.SNAKE;
