+++ source/core/src/main/com/csse3200/game/screens/GameOverLoseScreen.java
@@ -1,17 +1,13 @@
 package com.csse3200.game.screens;
 
-import com.badlogic.gdx.ScreenAdapter;
+
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.csse3200.game.GdxGame;
-import com.csse3200.game.components.gameover.GameOverLoseDisplay;
 import com.csse3200.game.components.gameover.GameOverActions;
+import com.csse3200.game.components.gameover.GameOverLoseDisplay;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.EntityService;
-import com.csse3200.game.entities.factories.RenderFactory;
 import com.csse3200.game.input.InputDecorator;
 import com.csse3200.game.input.InputService;
-import com.csse3200.game.rendering.RenderService;
-import com.csse3200.game.rendering.Renderer;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceLocator;
 import org.slf4j.Logger;
@@ -20,59 +16,22 @@ import org.slf4j.LoggerFactory;
 /**
  * The game screen for losing the game.
  */
-public class GameOverLoseScreen extends ScreenAdapter {
+public class GameOverLoseScreen extends ResizableScreen {
     private static final Logger logger = LoggerFactory.getLogger(GameOverLoseScreen.class);
     private final GdxGame game;
-    private final Renderer renderer;
     private static final String[] mainMenuTextures = {"images/box_boy_title.png"};
 
     public GameOverLoseScreen(GdxGame game) {
+        super();
+        
         this.game = game;
 
         logger.debug("Initialising game over win screen services");
         ServiceLocator.registerInputService(new InputService());
         ServiceLocator.registerResourceService(new ResourceService());
-        ServiceLocator.registerEntityService(new EntityService());
-        ServiceLocator.registerRenderService(new RenderService());
-
-        renderer = RenderFactory.createRenderer();
 
         loadAssets();
         createUI();
-    }
-
-    @Override
-    public void render(float delta) {
-        ServiceLocator.getEntityService().update();
-        renderer.render();
-    }
-
-    @Override
-    public void resize(int width, int height) {
-        renderer.resize(width, height);
-        logger.trace("Resized renderer: ({} x {})", width, height);
-    }
-
-    @Override
-    public void pause() {
-        logger.info("Game paused");
-    }
-
-    @Override
-    public void resume() {
-        logger.info("Game resumed");
-    }
-
-    @Override
-    public void dispose() {
-        logger.debug("Disposing main menu screen");
-
-        renderer.dispose();
-        unloadAssets();
-        ServiceLocator.getRenderService().dispose();
-        ServiceLocator.getEntityService().dispose();
-
-        ServiceLocator.clear();
     }
 
     private void loadAssets() {
