+++ source/core/src/main/com/csse3200/game/screens/LoadingScreen.java
@@ -1,17 +1,12 @@
 package com.csse3200.game.screens;
 
-import com.badlogic.gdx.ScreenAdapter;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.csse3200.game.GdxGame;
 import com.csse3200.game.components.animal.BackgroundImage;
 import com.csse3200.game.components.loading.LoadingDisplay;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.EntityService;
-import com.csse3200.game.entities.factories.RenderFactory;
 import com.csse3200.game.input.InputDecorator;
 import com.csse3200.game.input.InputService;
-import com.csse3200.game.rendering.RenderService;
-import com.csse3200.game.rendering.Renderer;
 import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceLocator;
@@ -19,52 +14,34 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /** The game screen containing the settings. */
-public class LoadingScreen extends ScreenAdapter {
+public class LoadingScreen extends ResizableScreen {
   private static final Logger logger = LoggerFactory.getLogger(LoadingScreen.class);
 
   private final GdxGame game;
-  private final Renderer renderer;
   private LoadingDisplay loadingDisplay;
 
   public LoadingScreen(GdxGame game) {
+    super();
+    
     this.game = game;
 
     logger.debug("Initialising loading screen services");
     ServiceLocator.registerInputService(new InputService());
     ServiceLocator.registerResourceService(new ResourceService());
-    ServiceLocator.registerEntityService(new EntityService());
-    ServiceLocator.registerRenderService(new RenderService());
     ServiceLocator.registerTimeSource(new GameTime());
 
-    renderer = RenderFactory.createRenderer();
     renderer.getCamera().getEntity().setPosition(5f, 5f);
 
     createUI();
   }
   @Override
   public void render(float delta) {
-    ServiceLocator.getEntityService().update();
-    renderer.render();
+    super.render(delta);
     if (loadingDisplay.isLoadingFinished()) {
       game.setScreen(GdxGame.ScreenType.MAIN_GAME);
     }
-
   }
-
-  @Override
-  public void resize(int width, int height) {
-    renderer.resize(width, height);
-  }
-
-  @Override
-  public void dispose() {
-    renderer.dispose();
-    ServiceLocator.getRenderService().dispose();
-    ServiceLocator.getEntityService().dispose();
-
-    ServiceLocator.clear();
-  }
-
+  
   /**
    * Creates the setting screen's ui including components for rendering ui elements to the screen
    * and capturing and handling ui input.
