+++ source/core/src/main/com/csse3200/game/screens/QuickTimeEventScreen.java
@@ -1,20 +1,15 @@
 package com.csse3200.game.screens;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.ScreenAdapter;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.utils.viewport.ExtendViewport;
 import com.csse3200.game.GdxGame;
 import com.csse3200.game.components.combat.quicktimeevent.QuickTimeEventActions;
 import com.csse3200.game.components.combat.quicktimeevent.QuickTimeEventDisplay;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.EntityService;
-import com.csse3200.game.entities.factories.RenderFactory;
 import com.csse3200.game.input.InputComponent;
 import com.csse3200.game.input.InputDecorator;
 import com.csse3200.game.input.InputService;
-import com.csse3200.game.rendering.RenderService;
-import com.csse3200.game.rendering.Renderer;
 import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceLocator;
@@ -25,10 +20,9 @@ import org.slf4j.LoggerFactory;
  * A screen dedicated to practicing quick-time events
  * used in main-game combat
  */
-public class QuickTimeEventScreen extends ScreenAdapter {
+public class QuickTimeEventScreen extends ResizableScreen {
     private static final Logger logger = LoggerFactory.getLogger(QuickTimeEventScreen.class);
     private final GdxGame game;
-    private final Renderer renderer;
 
     public QuickTimeEventScreen(GdxGame game) {
         this.game = game;
@@ -37,50 +31,21 @@ public class QuickTimeEventScreen extends ScreenAdapter {
         ServiceLocator.registerTimeSource(new GameTime());
         ServiceLocator.registerInputService(new InputService());
         ServiceLocator.registerResourceService(new ResourceService());
-        ServiceLocator.registerEntityService(new EntityService());
-        ServiceLocator.registerRenderService(new RenderService());
 
-        // Rendering
-        renderer = RenderFactory.createRenderer();
         renderer.getStage().setViewport(
-                new ExtendViewport(Gdx.graphics.getWidth(), Gdx.graphics.getHeight())
-        );
+                new ExtendViewport(Gdx.graphics.getWidth(), Gdx.graphics.getHeight()));
+        
         loadAssets();
         createUI();
     }
-
-    @Override
-    public void render(float delta) {
-        ServiceLocator.getEntityService().update();
-        renderer.render();
-    }
-
-    @Override
-    public void resize(int width, int height) {
-        renderer.resize(width, height);
-        logger.trace("Resized renderer: ({} x {})", width, height);
-    }
-
-    @Override
-    public void pause() {
-        logger.info("Game paused");
-    }
-
-    @Override
-    public void resume() {
-        logger.info("Game resumed");
-    }
-
+    
     @Override
     public void dispose() {
         logger.debug("Disposing QuickTimeEvent screen");
 
-        renderer.dispose();
         unloadAssets();
-        ServiceLocator.getRenderService().dispose();
-        ServiceLocator.getEntityService().dispose();
-
-        ServiceLocator.clear();
+        
+        super.dispose();
     }
 
     private void loadAssets() {
